Seismic sense shader
We made a shader that draws rings radiating from an impulse point (and some useful properties for ring appearance). Right now we have it set up so this impulse appears wherever you click. For now, you can only have one impulse point and collisions between objects are not considered for the propogation of the rings. We added some temporary models to our scene to demonstrate the shader.