It kinda works a little bit
Each thing with a collider and meshrenderer gets an impulse manager. The impulse managers keep track of which impulses each thing should be showing.
The GlobalImpulse class manages ImpulseManagers in groups. When they collide, they go into the same group. When collision exits, new groups are formed based on what is still touching.
When an impulse on something in a group reaches the point where it would start on another thing in the group that it's not already on, the GlobalImpulse class puts it on the other thing. The issue is that this requires Physics.ClosestPoint and it's very computationally expensive and complains a lot about meshes that aren't convex.
And also it flashes a lot and I'm really not sure why.